@@ -1,28 +1,26 @@
-source 'https://rubygems.org'
-ruby '2.4.1'
+source "https://rubygems.org"
+ruby "2.5.3"
 
-gem 'breakpoint',             '~> 2.5.0'
-gem 'builder',                '~> 3.2.2'
-gem 'html-proofer',           '~> 3.9'
-gem 'middleman',              '~> 3.4.1'
-gem 'middleman-autoprefixer', '~> 2.6.2'
-gem 'middleman-livereload',   '~> 3.4.5'
-gem 'middleman-minify-html',  '~> 3.4.1'
-gem 'middleman-syntax',       '~> 2.0.0'
-gem 'modular-scale',          '~> 2.1.1'
-gem 'nokogiri',               '~> 1.6'
-gem 'normalize-rails',        '~> 3.0.3', :require => false
-gem 'puma'
-gem 'rack'
-gem 'rack-contrib'
-gem 'rack-rewrite'
-gem 'rack-ssl',               '~> 1.4'
-gem 'rake'
-gem 'redcarpet',              '~> 3.3.3'
-gem 'semantic',               '~> 1.6'
-gem 'susy',                   '~> 2.2.9'
-gem 'therubyracer',           '~> 0.12.3'
-gem 'typogruby',              '~> 1.0.18'
-gem 'tzinfo-data',            '~> 1.2015.7', platforms: [:mswin, :mingw, :jruby]
-gem 'wdm',                    '~> 0.1.1',    platforms: [:mswin, :mingw]
-gem 'yard',                   '~> 0.9'
+gem "builder",                "~> 3.2.2"
+gem "html-proofer",           "~> 3.10"
+gem "middleman",              "~> 3.4.1"
+gem "middleman-autoprefixer", "~> 2.6.2"
+gem "middleman-livereload",   "~> 3.4.5"
+gem "middleman-minify-html",  "~> 3.4.1"
+gem "middleman-syntax",       "~> 2.0.0"
+gem "modular-scale",          "~> 2.1.1"
+gem "nokogiri",               "~> 1.6"
+gem "puma"
+gem "rack"
+gem "rack-contrib"
+gem "rack-rewrite"
+gem "rack-ssl",               "~> 1.4"
+gem "rack-conditional",       "~> 0.3"
+gem "rake"
+gem "redcarpet",              "~> 3.3.3"
+gem "semantic",               "~> 1.6"
+gem "therubyracer",           "~> 0.12.3"
+gem "typogruby",              "~> 1.0.18"
+gem "tzinfo-data",            "~> 1.2015.7", platforms: [:mswin, :mingw, :jruby]
+gem "wdm",                    "~> 0.1.1",    platforms: [:mswin, :mingw]
+gem "yard",                   "~> 0.9"
