@@ -1,5 +1,5 @@
 $(function() {
-  $('.anchors a[href*=#]:not([href=#])').click(function() {
+  $('a[href*=#]:not([href=#])').click(function() {
     if (location.pathname.replace(/^\//,'') == this.pathname.replace(/^\//,'')
         || location.hostname == this.hostname) {
 
